+++ CHANGELOG.md
@@ -4,6 +4,10 @@ We follow [Semantic Versions](https://semver.org/).
 
 ## unreleased
 
+- Add db invocations
+- Add k8s invocations
+- Add db_k8s invocations
+
 ## 0.4.1
 
 - Add `FastAPISettings` to imports